@@ -4,45 +4,39 @@
   "description": "",
   "main": "Server/server.js",
   "scripts": {
-    "dev": "set NODE_ENV=development&& start /b cmd /c mongod & tsc-watch --onSuccess \"node ./Server/backend/server.js\"",
+    "dev": "set NODE_ENV=development& tsc-watch --onSuccess \"node ./Server/backend/server.js\"",
     "predev": "shx cp ./node_modules/socket.io/client-dist/socket.io.js ./Server/public/javascripts/socket.io.js",
-    "start": "set NODE_ENV=production&& start /b cmd /c mongod & node ./Server/backend/server.js && taskkill /f /im mongod.exe"
+    "start": "set NODE_ENV=production& node ./Server/backend/server.js "
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "bcrypt": "^5.0.0",
-    "body-parser": "^1.19.0",
-    "connect-flash": "^0.1.1",
-    "connect-mongo": "^4.6.0",
-    "dotenv": "^16.0.3",
-    "express": "^4.17.1",
-    "express-ejs-layouts": "^2.5.0",
-    "express-session": "^1.17.1",
-    "formidable": "^2.0.1",
-    "mongoose": "^6.1.8",
-    "mongoose-validator": "^2.1.0",
-    "nodemon": "^2.0.6",
-    "passport": "^0.4.1",
+    "bcrypt": "^5.1.1",
+    "better-sqlite3": "^11.7.0",
+    "better-sqlite3-session-store": "^0.1.0",
+    "dotenv": "^16.4.7",
+    "express": "^4.21.2",
+    "express-session": "^1.18.1",
+    "formidable": "^3.5.2",
+    "md5": "^2.3.0",
+    "passport": "^0.7.0",
     "passport-local": "^1.0.0",
-    "pm2": "^5.2.2",
-    "socket.io": "^4.5.3",
-    "websocket": "^1.0.33",
-    "yargs": "^17.6.0"
+    "socket.io": "^4.8.1",
+    "websocket": "^1.0.35",
+    "yargs": "^17.7.2"
   },
   "devDependencies": {
-    "@types/bcrypt": "^3.0.0",
-    "@types/express": "^4.17.9",
-    "@types/express-session": "^1.17.3",
-    "@types/formidable": "^2.0.1",
-    "@types/passport": "^1.0.4",
-    "@types/ps-node": "^0.1.1",
-    "@types/ws": "^7.4.0",
-    "@types/yargs": "^17.0.13",
-    "jshint": "^2.13.4",
+    "@types/bcrypt": "^5.0.2",
+    "@types/better-sqlite3": "^7.6.12",
+    "@types/express": "^5.0.0",
+    "@types/express-session": "^1.18.1",
+    "@types/formidable": "^3.4.5",
+    "@types/md5": "^2.3.5",
+    "@types/passport": "^1.0.17",
+    "@types/yargs": "^17.0.33",
     "shx": "^0.3.4",
-    "tsc-watch": "^4.6.0",
+    "tsc-watch": "^6.2.1",
     "tslint": "^6.1.3",
-    "typescript": "^4.1.3"
+    "typescript": "^5.7.2"
   }
 }
